Modify dummy target details

+++ src/include/Global.h
@@ -9,7 +9,7 @@
 typedef unsigned int tiles_t;
 #endif
 #ifdef TARGET_DUMMY
-typedef uint8_t tiles_t;
+typedef uint16_t tiles_t;
 #endif
 #ifdef TARGET_GBA
 #include <gba_types.h>
